update：post action UI
