+++ src/pages/Topic/PostItem/Header.tsx
@@ -1,10 +1,9 @@
-import React, { useState } from 'react'
+import React from 'react'
 
 import { css } from 'emotion'
 import { navigate } from '@reach/router'
 
-import { CardHeader, Avatar, IconButton, Menu, MenuItem } from '@material-ui/core'
-import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
+import { CardHeader, Avatar, IconButton } from '@material-ui/core'
 import { ClassNameMap } from '@material-ui/core/styles/withStyles'
 import { IPost, IUser } from '@cc98/api'
 
@@ -12,8 +11,6 @@ interface Props {
   classes: ClassNameMap
   postInfo: IPost
   userInfo: IUser | null
-  isTrace: boolean
-  openDialog: (info: IPost) => void
 }
 
 const cursorStyle = css`
@@ -25,9 +22,7 @@ const postOptionStyle = css`
 `
 
 export default (props: Props) => {
-  const { classes, postInfo, userInfo, isTrace, openDialog } = props
-  const [anchorEl, setAnchorEl] = useState(null)
-  const open = Boolean(anchorEl)
+  const { classes, postInfo, userInfo } = props
 
   return (
     <CardHeader
@@ -54,57 +49,6 @@ export default (props: Props) => {
             {postInfo.isHot ? '热' : `${postInfo.floor}`}
           </Avatar>
         </IconButton>,
-        // tslint:disable-next-line:ter-indent
-        <div key="options" className={postOptionStyle}>
-          <IconButton
-            key="option"
-            classes={{ root: classes.iconRoot }}
-            aria-label="More"
-            aria-owns={open ? 'long-menu' : undefined}
-            aria-haspopup="true"
-            // tslint:disable-next-line:no-any
-            onClick={(e: any) => setAnchorEl(e.currentTarget)}
-          >
-            <ExpandMoreIcon fontSize="small" />
-          </IconButton>
-          <Menu
-            id="long-menu"
-            anchorEl={anchorEl}
-            open={open}
-            onClose={() => setAnchorEl(null)}
-            PaperProps={{
-              style: {
-                maxHeight: 48 * 4.5,
-                width: '5rem',
-              },
-            }}
-          >
-            {['评分', isTrace ? '返回' : '追踪', '编辑'].map(option => (
-              <MenuItem
-                key={option}
-                onClick={() => {
-                  if (option === '追踪') {
-                    if (!postInfo.isAnonymous) {
-                      navigate(`/topic/${postInfo.topicId}/trace/${postInfo.userId}`)
-                    } else {
-                      navigate(`/topic/${postInfo.topicId}/anonymous/trace/${postInfo.id}`)
-                    }
-                  } else if (option === '返回') {
-                    navigate(`/topic/${postInfo.topicId}`)
-                  } else if (option === '编辑') {
-                    // TODO:
-                  } else if (option === '评分') {
-                    openDialog(postInfo)
-                  }
-                  setAnchorEl(null)
-                }}
-                classes={{ root: classes.menuItemRoot }}
-              >
-                {option}
-              </MenuItem>
-            ))}
-          </Menu>
-        </div>,
       ]}
     />
   )
