+++ src/pages/Topic/PostItem/index.tsx
@@ -4,15 +4,12 @@ import { css, cx } from 'emotion'
 
 import UBB from '@cc98/ubb-react'
 
-import Utils from './PostUtils'
-
 import {
   Button,
   Card,
   CardActions,
   CardContent,
   Collapse,
-  Divider,
   IconButton,
   Typography,
 } from '@material-ui/core'
@@ -20,12 +17,10 @@ import {
 import { StyleRules, withStyles } from '@material-ui/core/styles'
 import { ClassNameMap } from '@material-ui/core/styles/withStyles'
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore'
-import Quote from '@material-ui/icons/RotateLeft'
-import DislikeIcon from '@material-ui/icons/ThumbDown'
-import LikeIcon from '@material-ui/icons/ThumbUp'
 
 import Header from './Header'
 import Award from './Award'
+import Actions from './Actions'
 
 import resolveMarkdown from './resolveMarkdown'
 import { IPost, IUser } from '@cc98/api'
@@ -68,23 +63,6 @@ interface Props {
   setPosts: <T extends Partial<IPost>>(postId: number, postUpdate: T) => void
 }
 
-const likeStateMap = ['none', 'like', 'dislike']
-const likeButton = {
-  clicked: css`
-    color: #dd5e5c;
-  `,
-  unclicked: css`
-    color: inherit;
-  `,
-}
-const dislikeButton = {
-  clicked: css`
-    color: #6464ff;
-  `,
-  unclicked: css`
-    color: inherit;
-  `,
-}
 const styles: StyleRules = {
   actionsRoot: {
     display: 'flex',
@@ -92,10 +70,7 @@ const styles: StyleRules = {
     height: '2rem',
   },
   action: {
-    flexGrow: 1,
-    '&:hover': {
-      backgroundColor: '#fff',
-    },
+    width: '4rem',
   },
   hr: {
     border: '#555 solid thin',
@@ -107,6 +82,7 @@ const styles: StyleRules = {
   },
   iconRoot: {
     padding: '5px',
+    marginLeft: 'auto',
   },
   menuItemRoot: {
     width: '3rem',
@@ -168,13 +144,8 @@ export default withStyles(styles)((props: Props) => {
 
   return (
     <Card square elevation={0} className={root}>
-      <Header
-        postInfo={postInfo}
-        userInfo={userInfo}
-        isTrace={isTrace}
-        classes={classes}
-        openDialog={openDialog}
-      />
+      <Header postInfo={postInfo} userInfo={userInfo} classes={classes} />
+
       <CardContent>
         <Typography
           classes={{ root: classes.typographyRoot }}
@@ -185,68 +156,15 @@ export default withStyles(styles)((props: Props) => {
         </Typography>
       </CardContent>
 
-      <CardActions classes={{ root: classes.actionsRoot }}>
-        <IconButton
-          classes={{ root: classes.action }}
-          disableRipple={true}
-          onClick={async () => {
-            const res = await Utils.dislike(postInfo.id)
-            setPosts(postInfo.id, res)
-          }}
-        >
-          <DislikeIcon
-            fontSize="small"
-            className={
-              likeStateMap[postInfo.likeState] === 'dislike'
-                ? dislikeButton.clicked
-                : dislikeButton.unclicked
-            }
-          />
-          <span
-            key="dislikeIcon"
-            style={{ fontSize: '0.9rem', marginLeft: '0.875rem', opacity: 0.54 }}
-          >
-            {postInfo.dislikeCount}
-          </span>
-        </IconButton>
-        <Divider classes={{ root: classes.hr }} />
-        <IconButton
-          classes={{ root: classes.action }}
-          disableRipple={true}
-          onClick={async () => {
-            const content = Utils.quote(postInfo)
-            setQuote(content)
-          }}
-        >
-          <Quote fontSize="small" />
-        </IconButton>
-        <Divider classes={{ root: classes.hr }} />
+      <Actions
+        postInfo={postInfo}
+        setPosts={setPosts}
+        setQuote={setQuote}
+        openDialog={openDialog}
+        isTrace={isTrace}
+        classes={classes}
+      />
 
-        <IconButton
-          classes={{ root: classes.action }}
-          disableRipple={true}
-          disableTouchRipple={true}
-          onClick={async () => {
-            const res = await Utils.like(postInfo.id)
-            setPosts(postInfo.id, res)
-          }}
-        >
-          <LikeIcon
-            fontSize="small"
-            className={
-              likeStateMap[postInfo.likeState] === 'like'
-                ? likeButton.clicked
-                : likeButton.unclicked
-            }
-          />
-          <span
-            key="likeIcon"
-            style={{ fontSize: '0.9rem', marginLeft: '0.875rem', opacity: 0.54 }}
-          >
-            {postInfo.likeCount}
-          </span>
-        </IconButton>
-      </CardActions>
       {postInfo.awards.length > 5 && (
         <CardActions classes={{ root: classes.awardAction }}>
           <Button classes={{ root: classes.expandButton }} onClick={onExpandClick}>
